Camera Lerping to Controlled unit
CameraController script now looks at & positions itself relative to a focus point. This point is currently just the selected squad member's position. 
Added OnSwitchSquaddie event to stSquadManager class. This event is called each time the active squad member is changed.
CameraController now has an event listener for the above event. When the player switches character, the event is fired which puts the camera into lerping mode. This allows for smooth transitions from previous to next squad member. 
CameraController now has a float _switchCharacterTime. This is the time in seconds it will take for the camera to switch to the new squad members location.
CameraController's rotation is now taken into account when positioning the camera.