Force Text Serialization & Aim Script work
Changed Asset Serialization to Force Text mode for proper source control integration.
Renamed CameraController script's _centerPoint variable to _focusPoint.
PlayerAim script is now able to find the first object under the mouse position which has a tag specified in the aimTags array.
Added a SwitchTo function to stSquadManager class which allows a character to be directly switched to via their index in the squad array.