Event fixes
Slightly optimized SwitchSquadMemberHandler script, now directly calls switch function without calling a second function or debug logging.
stSquadManager class now fires the OnSwitchSquaddie event to properly deselect the current squad member when changing the list of selectable members. 
Added SafeFireOnSwitchSquaddie method to properly verify there are registered listeners for the event before firing.
SquaddieSwitchController now subscribes to the OnSwitchSquaddie event in stSquadManager. This allows the script to call it's Select or Deselect method when the event is fired, without any other scripts needing to access & call these functions directly.
PlayerAim script now records the point the player is aiming at.
Began implementation for GetAimViewportOffset function in CameraController script. When finished, this will allow the camera to slightly offset from the player's position in the direction of the mouse on the screen. (See alien swarm game for the basic idea).