Camera Features + Closer to smooth camera movement
WORK IN PROGRESS! Lerping functionality currently does not work properly!

Refactored some code in the CameraController script to be neater.
Completely rewrote how the GetAimViewportOffset function works. It is now simply called GetAimOffset, as it no longer uses the viewport.
New variable _aimOffsetDistance was added to control how much the aim point influences the camera.
New variable _pitch was added to allow the user to specify the angle at which the camera views the focus point.
New variable _distance was added to control the distance between the camera and the focus point.