Complete rewrite for CameraController
Completely rewrote the CameraController class to now operate much more smoothly, and with neater code!
The CameraController class is now designed to be attached to the camera itself, rather than a GameManager gameobject.
Overhauled the variables from the previous script to be much neater and more obvious as to what they are for.
The script now simply calculates a look point, which is smoothly lerped. Position is no longer lerped at the same time, to prevent weird wobbly & sickening movement. Position is now simply set to the lerped look-point, plus an offset.
Added "start transform" variables which allow some data to be overridden at start if a boolean is true.
Added farDistance and farSpeed variables which are used in conjunction with the AnimationCurve to calculate the desired speed at any given distance from the destination point.
Cleaned up the inspector front-end, and added Headers and many useful tooltips.