Auto stash before merge of "WIP" and "origin/WIP"
Tidied up & commented several scripts.
Renamed SquaddieSwitchController script to SquaddieController. This script should serve more purposes than just switching between characters.
Added temporary test script called LaserTest to simply draw a line to where the player is aiming. This will later be replaced by a better looking laser pointer.
Created a simple LogOnce class which allows a specified message to be logged a single time. This removes the need for individual scripts to track a boolean for whether or not they have already logged a warning, etc. 
Changed _aimPoint to _aimMousePoint in PlayerAim script. This is later intended to be different to the actual aimPoint, which may need to be raised or lowered based on actual character line of sight. However, some scripts may still need info on where the mouse is hovered, eg. for the CameraController's aim offset.
stSquadManager class no longer returns a SquaddieController in the Switch function, as this is not needed.