Utility Functions & More work on player aim script
Created a class for Utility functions, under the namespace JakePerry.
The Utils class currently includes a function called IsBetween for easily determining if a given value is between a specified min and max. It also features an ApplyTagMask function for Transform, Collider, and RaycastHit arrays. This simply iterates through each object in the array, and removes any that do not have a tag found in the tag array.
Also created an Extensions class under the namespace JakePerry.
The extensions class has extension methods for Vector types to determine if a vector's components are all in a given range, as well as functions to get a Vector of the minimum or maximum components of two Vectors.
The extension class also has a function to find the top level parent of a transform.
Also wrote a function to get the full bounds of a colider & all child object colliders. The function is not currently working correctly, however.

PlayerAim script now aims to check for a more optimal aim point. The functionality is still a work in progress, but a _aimPoint variable will now be set to the optimal point. In most cases, this will simply be the point under the mouse for now.