PostProc stack asset included, more util functions + aim changes
Added the Post Processing Stack asset to the project.
Added Utility functions for filtering an array of Transform, RaycastHit or Collider objects, called GetChildrenOf. The function returns an array of elements in the specified collection that have a transform that is a child of the specified parent transform.
Changed some default values in the CameraController script to feel better when smoothly moving around.
stSquadManager will now only allow the player to switch squad members if the stSettings class allows it. (The settings class has a _canSwitchSquadMembers variable).