SmartAim now works, fixed several scripts
PlayerAim script's SmartAim functionality now works as intended. The _smartAimIterations variable, available in the inspector, determines how many increments rays should be targeted at across the bounds of the object under the mouse. These are iterated over in all 3 directions, until a point is found which is visible to the character.
All newly added work in PlayerAim script makes use of RaycastNonAlloc, to prevent too much excess memory from being allocated each frame. A raycastHit variable and array have been added to the script to hold this information. 
Several public functions across a few scripts have been turned into private functions, with the SerializeField attribute so they are still accessible by the inspector. This will prevent variables from being accessible to other scripts that shouldnt touch them.
Added a generic extension function for arrays to check if the array contains a specified object.
Added a utility function to get first instance from an array with a tag matching a specified string array. This is overloaded for Transform, Collider and RaycastHit.