Smart Aim optimization
PlayerAim script's smart aim functionality will now track all valid points when finding an optimal point, rather than returning the valid point closest to the transform's origin. The list of valid points is then searched, and the point closest to the original mouse point is used. This helps to achieve a smoother feel when aiming.

Directions can now be ignored when checking points within bounding box, if the character's angle to the transform is close enough to that direction. This can drastically reduce the number or raycast checks performed (eg from ~130 to ~40 when iterations is set to 5). 
The ray to the bounds center is no longer checked for each iteration. This will now only be checked a single time, removing needless overhead.
Removed unneeded line of code @ line 236.