Smart Aim optimization
Smart Aim functionality in PlayerAim script is now even more optimized!
Rather than tracking a list of all valid points and comparing distance at the end, the script now tracks a single vector and float for distance. The distance to the current offset point is checked before any raycasting or sorting is done, which dramatically cuts down on the number of tests.