Utility AI System started.
Added function to CameraController to get the forward vector along horizontal.

Started implementation for Utility AI system.
Added AI properties. Base property is abstract, and contains an abstract c# property for returning a normalized value. 
Added property classes for bool, float, int for use with AI system.
Added AI Behaviour base class. The implementation of this class may need to be changed, but currently serves as a base for all behaviours. 
File also contains struct for considerations used by behaviour class. The consideration has a property name and animationCurve which is used for weighing priority. The implementation may need to be changed to store a reference to a property instance, however this will depend on how everything is implemented with the Unity inspector.
Currently, the behaviour base class has a list of considerations. This should be iterated through to find a priority weight for the tied action.
Added AI Agent class derived from monobehaviour. This agent script has a list of behaviours. Each update call, every behaviour will be weighed up and the top priority behaviour will be followed.