AI Property heirarchy changed to work better with serialization
Properties for use with Utility AI system are now a single class which stores all variable types. This change has been made due to the way Unity serializes classes. Abstract classes were not able to be serialized, and any extra data on subclasses are not serialized properly either.
Property class is now named uaiProperty. A custom property drawer still needs to be implemented properly to draw based on the property type.