Properties & now finally working & drawing properly
uaiProperty is now a struct rather than a class.
uaiAgent now uses the "nullable" modifier '?' for some functions to be able to return a null struct reference. This is useful for when no property with a given name is found, etc, so null is returned.
All properties are now drawn directly via the draw callback function on the reorderable list in the uaiAgent editor.

Known Issue: While the scene is playing, property values do not seem to be modifiable. It's possible they are for some reason being reset to their previous value each frame, but after disabling the custom drawer, the issue still persisted. Will have to fix this soon.