Method for finding proper transform parent
Added extension function for Collider called TopParentMatchingTag . The function will ascend the heirarchy to find the top-level collider attached to the this-collider's heirarchy which has the same tag as the collider.
PlayerAim now uses the added function to find an optimized point on the object being aimed at. This is useful for finding the parent object for the enemy (when aiming at legs, etc), without ascending too high and getting the bounds of an empty gameObject which may contain more than one enemy.