Squad-Switch delays, Aim LayerMask
PlayerAim script now has a layermask which is used for all raycasting.
SpawnNode's OnDestroy method should now be called correctly.
SetSquadReferences script now checks for null references and handles them properly.
SquaddieController script will now invoke the selection of a squad member after a delay. This can be set to select as soon as the camera switches to the new member.