AudioPlayer inspector fix + Gun shot-fired event
Fixed the AudioPlayer inspector so that changes to the reorderable list are saved across frames.
Added an event to the Gun script which is raised when the gun is fired. 
Added a listener script for this event which is capable of playing a sound when the gun is fired.

Changed Gun script's maximum value for fire rate (in RPM) to 900 rather than 300 as it felt too low.
Updated player prefab to play sound when shooting.