Added toggle for time-scale pitch per AudioPlayer
Added a boolean variable timeScaleInfluencesPlayback to the AudioPlayer script. This is queried by the manager before affecting the pitch of the source.
This addition will allow music and other sounds to be played even when the game is paused, etc.

AudioPlayer's channel variable is now private, with a public accessor.